pyproject.toml
[project]
name = "glowcast"
version = "1.0.0"
description = "Beauty & skincare supply chain AI platform — demand forecasting, causal inference, A/B testing, and MLOps for 5,000 SKUs across 12 fulfillment centers."
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    # Core data
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "scikit-learn>=1.4.0",
    "pandera>=0.18.0",
    # Forecasting
    "lightgbm>=4.0.0",
    "xgboost>=2.0.0",
    # Config & logging
    "structlog>=23.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "hypothesis>=6.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]
forecast = [
    "statsforecast>=1.7.0",
    "hierarchicalforecast>=0.4.0",
]
mlops = [
    "evidently>=0.4.0",
    "mlflow>=2.10.0",
]
causal = [
    "dowhy>=0.11.0",
    "causalml>=0.15.0",
    "econml>=0.15.0",
]
explain = [
    "shap>=0.45.0",
    "lime>=0.2.0",
]
chronos = ["chronos-forecasting>=1.3.0"]
all = [
    "glowcast[forecast,mlops,causal,explain,chronos]",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 130
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "S", "B"]
ignore = ["S101", "S603", "S607", "B008", "E402", "S311", "B905", "B007"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
